requirements.txt
beautifulsoup4==4.6.0
lxml==4.2.1
Pillow==5.1.0
